pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssl-assessment"
version = "0.1.0"
description = "HF-only SAP x QualiSpeech assessment pipeline"
requires-python = ">=3.10"
dependencies = [
  "huggingface_hub>=0.30.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "pyyaml>=6.0.0",
  "scipy>=1.12.0",
  "tabulate>=0.9.0",
  "torch>=2.3.0",
  "torchaudio>=2.3.0",
  "transformers>=4.50.0",
  "pytest>=8.0.0",
]

[tool.setuptools]
packages = [
  "src",
  "src.analysis",
  "src.cli",
  "src.data",
  "src.eval",
  "src.models",
  "src.plots",
  "src.samplers",
  "src.tables",
  "src.tasks",
  "src.trainers",
  "src.utils",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
